issue #62 feature : GetPlanList -> PlanState Mapper 추가
