--- client/app/src/main/java/com/example/witt/domain/model/plan/GetPlanListModel.kt
@@ -1,11 +1,0 @@
-package com.example.witt.domain.model.plan
-
-import java.time.LocalDateTime
-
-data class GetPlanListModel (
-    val PlanId: String,
-    val StartDate: LocalDateTime,
-    val EndDate: LocalDateTime,
-    val Name: String,
-    val Region: String
-        )
